order detail service

+++ src/main/java/com/example/shopapp/model/OrderDetail.java
@@ -30,7 +30,7 @@ public class OrderDetail {
     private int numberOfProducts;
 
     @Column(name = "total_money", nullable = false)
-    private int totalMoney;
+    private Float totalMoney;
 
     @Column(name = "color", nullable = false)
     private String color;
